+++ src/lib.rs
@@ -1,3 +1,6 @@
 pub mod btc;
 pub mod mempool;
 pub(crate) mod util;
+
+#[cfg(test)]
+mod test;
