+++ src/mempool.rs
@@ -106,68 +106,3 @@ impl MempoolFilterReader {
         self.reader.match_all(reader, query)
     }
 }
-
-
-#[cfg(test)]
-mod tests {
-
-use bitcoin::Block;
-use bitcoin::consensus::deserialize;
-use bitcoin::hashes::hex::FromHex;
-use bitcoin::Transaction;
-use std::collections::HashMap;
-use std::fs;
-use std::io;
-use std::io::BufRead;
-use super::*;
-
-#[test]
-    fn block_00000000000000000007fc62780dee62d79ba02e7d325d7503e80c4da8b16b72() {
-        let k0 = u64::from_le_bytes(*b"qwertyui");
-        let k1 = u64::from_le_bytes(*b"opasdfgh");
-        let block = load_block("./test/block1");
-        let txmap = make_inputs_map(load_txs("./test/block1-txs"));
-        let mut txs = block.txdata;
-        txs.remove(0); // remove coinbase
-        let txs2 = txs.clone();
-        let test_filter = ErgveinMempoolFilter::new_script_filter
-            (k0,k1, txs,
-                |o| if let Some(s) = txmap.get(o) {
-                    Ok(s.clone())
-                } else {
-                    Err(Error::UtxoMissing(o.clone()))
-                }).unwrap();
-        let tx = &txs2[11];
-        assert_eq!(test_filter.match_tx_outputs(k0,k1, &tx).unwrap(), true);
-    }
-
-    fn load_block(path: &str) -> Block {
-        let mut contents = fs::read_to_string(path).unwrap();
-        contents.pop();
-        deserialize(&Vec::from_hex(&contents).unwrap()).unwrap()
-    }
-
-    fn make_inputs_map(txs: Vec<Transaction>) -> HashMap<OutPoint, Script> {
-        let mut map = HashMap::new();
-        for tx in txs {
-            let mut out_point = OutPoint {
-                txid: tx.txid(),
-                vout: 0,
-            };
-            for (i, out) in tx.output.iter().enumerate() {
-                out_point.vout = i as u32;
-                map.insert(out_point.clone(), out.script_pubkey.clone());
-            }
-        }
-        map
-    }
-    fn load_txs(path: &str) -> Vec<Transaction> {
-        let mut res = vec![];
-        let file = std::fs::File::open(path).unwrap();
-        for line in io::BufReader::new(file).lines() {
-            let tx = deserialize(&Vec::from_hex(&line.unwrap()).unwrap()).unwrap();
-            res.push(tx);
-        }
-        res
-    }
-}
